% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLipoReport.R
\name{addValues}
\alias{addValues}
\title{add a serie of value to an existing lipoprotein  quantification report}
\usage{
addValues(lip, options = list())
}
\arguments{
\item{lip}{\itemize{
\item a lipoprotein report with new values
}}

\item{options}{\itemize{
\item a list of options (should be as existing report)
}}

\item{options$fold}{\itemize{
\item if true use design for fold change (default = FALSE)
}}

\item{options$scale}{\itemize{
\item if true center and scale (align) the output (default = FALSE)
}}
}
\value{
print a report with overlayed values
}
\description{
add a serie of value to an existing lipoprotein  quantification report
}
