% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRClassDefinitions.R
\docType{class}
\name{Analyte-class}
\alias{Analyte-class}
\title{An S4 class for Analyte}
\value{
a dataElement
}
\description{
An S4 class for Analyte
}
\section{Slots}{

\describe{
\item{\code{signals}}{a list of NMRSignal1D}

\item{\code{category}}{A string specifying a category for the compound. ie.e: 'metabolite', 'substrate source'}

\item{\code{name}}{A human readable name for the analyte}

\item{\code{inchiKey}}{A string used to identify a molecule}

\item{\code{diaID}}{A canonical molecule ID}

\item{\code{type}}{The Class Name. Used for moving between S4 and JSON}

\item{\code{id}}{(optional) Unique identifier}
}}

