% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcASCVD.R
\name{calcASCVD}
\alias{calcASCVD}
\title{function to calculate ASCVD risk}
\usage{
calcASCVD(
  Gender = c("F", "M"),
  Ethnicity_African = FALSE,
  HypertensionMeds = FALSE,
  Smoking = FALSE,
  DM = FALSE,
  Age = NULL,
  TPCH = NULL,
  HDCH = NULL,
  SBP = NULL
)
}
\arguments{
\item{Gender}{character either "F" female or "M" male}

\item{Ethnicity_African}{True or False}

\item{HypertensionMeds}{Hypertension Medication use (True or False)}

\item{Smoking}{Current Smoking Status (True or False)}

\item{DM}{Diabetes status (True or False)}

\item{Age}{numeric and must be between 40 to 79}

\item{TPCH}{numeric and must be 130 mg/dl and above}

\item{HDCH}{numeric and must be above 20 mg/dl}

\item{SBP}{numeric and must be above 90 mmHg}
}
\value{
numeric a ASCVD risk in percentage
}
\description{
function to calculate ASCVD risk
}
\examples{

risk <- calcASCVD(Gender = "M", Ethnicity_African = TRUE, HypertensionMeds = FALSE, Age = 55, TPCH = 213, HDCH = 50, SBP = 120)
}
\references{
https://pubmed.ncbi.nlm.nih.gov/24222018/

https://clincalc.com/Cardiology/ASCVD/PooledCohort.aspx
}
