% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.R
\name{cbind.dataElement}
\alias{cbind.dataElement}
\title{cbind dataElements}
\usage{
\method{cbind}{dataElement}(...)
}
\arguments{
\item{...}{a list of dataElements}
}
\value{
the information about dataElement
}
\description{
cbind dataElements
}
