% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDataElement.R
\name{check,dataElement-method}
\alias{check,dataElement-method}
\title{SS4 method to check dataElement}
\usage{
\S4method{check}{dataElement}(da)
}
\arguments{
\item{da}{dataElement object to be checked}
}
\value{
void
}
\description{
SS4 method to check dataElement
}
\examples{

check("dataElement")

}
