% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDataElement.R
\name{check}
\alias{check}
\title{S4 method to check dataElement
#'}
\usage{
check(da)
}
\arguments{
\item{da}{dataElement object to be checked}
}
\value{
void
}
\description{
S4 method to check dataElement
#'
}
\examples{

check("dataElement")

}
