% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRun.R
\name{checkRun}
\alias{checkRun}
\title{checkRun
collect information if the 1D, CPMG, PGPE has been run, Bruker Report (QC, LIPO, Quant, PACS) has been generated for each samples/ltr as planned or not.}
\usage{
checkRun(runID = "SPTr04", by = "row", ltr = "both")
}
\arguments{
\item{runID}{to link between the server and the expected}

\item{by}{the order of the EXPNO default is by row and alternative is "column"}

\item{ltr}{to consider ltr and/or sltr when calculating the EXPNO.
default is ltr = "both" assuming 4 ltr and 4 sltr per plate.
if ltr = "sltr" meaning the plate has 0 ltr but 4 sltr per plate
if ltr = "ltr" meaning the plate has 4 ltr and 0 sltr per plate
if ltr = "none" meaning 0 ltr and 0 sltr per plate}
}
\value{
tdf1

first the expected samples per plate is gather from the run Layout.
This must be matched with sample layout found in the configurationfile.
from "expected", combination of the "plateName", "expno", and the "sampleID" named "expname" should be unique.
then the dataPath information about the 1D noesy run in the server is extracted from loe.
from "dataPath", combination of the "plateName", "expno", and the "sampleID" named "check" should be unique.
by mergin those two by "expname" and "check" you have the new data frame called "tdf1" that capture which samples/ltrs has been run and which are the re-runs.

the next will be a qc report check.
match the previous tdf1 using the dataPath this time to see how many of the run has QC report and which are missing.
\%\%\% Note: only the sampleMatrixType of SER, PLA, and URI should have qc report \%\%\%
only if the qc report exists, lipo, quant, pacs report will also get checked.

once the 1D run has been checked then cpmg and the pgpe (zggppezf) run will be checked if exist.
for cpmg the EXPNO - 1 to match the 1D dataPath
for zggppezf, the EXPNO - 2 to match the 1D dataPath
}
\description{
#'
}
