% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDataElement.R
\docType{class}
\name{dataElement-class}
\alias{dataElement-class}
\title{An S4 class for dataElement}
\value{
a dataElement
}
\description{
An S4 class for dataElement
}
\section{Slots}{

\describe{
\item{\code{.Data}}{the data matrix}

\item{\code{varName}}{a vector containing the name of each variable}

\item{\code{obsDescr}}{a data.frame containing experimental
conditions and a field called sampleID that MUST be unique.}

\item{\code{method}}{a name for the method used to acquire the data}

\item{\code{type}}{type can be NMR, U-MS, T-MS, ANN, MRMS}
}}

