% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dduplicated}
\alias{dduplicated}
\title{a function to get duplicated pairs}
\usage{
dduplicated(vector)
}
\arguments{
\item{vector}{\itemize{
\item a vector
}}
}
\value{
a logical vector
}
\description{
a function to get duplicated pairs
}
