% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delist}
\alias{delist}
\title{a function to unlist but replacing empty values by NA to keep the length}
\usage{
delist(l, unname = TRUE)
}
\arguments{
\item{l}{\itemize{
\item a list
}}

\item{unname}{\itemize{
\item boolean switch to unname response (default = true)
}}
}
\value{
a vector
}
\description{
a function to unlist but replacing empty values by NA to keep the length
}
