% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendLipo.R
\name{extendLipo}
\alias{extendLipo}
\title{function to calculate total lipids, subfraction and the percentage}
\usage{
extendLipo(data)
}
\arguments{
\item{data}{\itemize{
\item lipoprotein daE or other format such as data.frame
}}
}
\value{
112 new column of either calc. (calculated values) or perc. (percentage)
}
\description{
function to calculate total lipids, subfraction and the percentage
}
\examples{
#data<-local(get(load("~lipo.daE")))
#extended_lipo<-extend_lipo(data)
}
