% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterWith.R
\name{filterWith,dataElement,logical-method}
\alias{filterWith,dataElement,logical-method}
\title{S4 method to filter a dataElement}
\usage{
\S4method{filterWith}{dataElement,logical}(da, fi)
}
\arguments{
\item{da}{first dataElement}

\item{fi}{the filter}
}
\value{
the reordered dataElement
}
\description{
S4 method to filter a dataElement
}
