% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRClassDefinitions.R
\name{fromVector,ANY-method}
\alias{fromVector,ANY-method}
\title{Introspect a data object and transform any matching structure into the
corresponding S4 object}
\usage{
\S4method{fromVector}{ANY}(input)
}
\arguments{
\item{input}{A data object to be parsed (list, array or S4)}
}
\value{
an object
}
\description{
Introspect a data object and transform any matching structure into the
corresponding S4 object
}
