% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\name{fusion}
\alias{fusion}
\title{S4 method melt dataset objects (S4)}
\usage{
fusion(v, ...)
}
\arguments{
\item{v}{visualization object}

\item{...}{allows to pass translate = TRUE for running into docker instance}
}
\value{
void
}
\description{
A visualize method for \emph{visualization} objects. If \emph{object@localServer} is TRUE, then this
function will trigger a local instance of \pkg{servr} server.
}
\examples{


}
