% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSION.R
\name{FUSION_ps}
\alias{FUSION_ps}
\title{FUSION_ps - Enrichment Analysis of sncRNA families in paired data samples using expression matrix}
\usage{
FUSION_ps(
  a,
  row_mean = getOption("row_mean", "0.1"),
  sncrna_family = c("tsrna", "rsrna", "ysrna", "mirna", "other"),
  padj_method = c("bonferroni", "BH"),
  order = c("G", "P")
)
}
\arguments{
\item{a}{a matrix with first and second column as "Sequence" (or ID) and "Annotation", respectively. Sequence (or ID) must be unique.  Rest of the columns provides RPM or expression values from different samples under study, such as first half of those belongs to the samples from Condition1 (such as control or from healthy tissue) and second half belongs to the samples from Condition2 (such as treated or infected tissue)}

\item{row_mean}{mean RPM (default 0.1) threshold to retain the sncRNA species (rows) in the matrix}

\item{sncrna_family}{list of sncRNA families to be analysis for expression analysis study. Use "tsrna", "rsrna", "ysrna", or "mirna" for tsRNAs, rsRNAs, ysRNAs, or mirna, respectively. Use "other" for 'pRNA,snRNA and snoRNA'. For all, use any letter or number, eg. "a","b", "c", 1, 2, 3. By default (i.e. if no option specified) it will search for tryRNAs (tsRNAs, rsRNAs, and ysRNAs)}

\item{padj_method}{adjustment methods for correcting p-values. Use either "bonferroni" or "BH", where BH stands for Benjamini & Hochberg. By default (i.e. if no option specified) it will run for "bonferroni".}

\item{order}{use either G or P for specifying the order of paired samples in the input matrix, where G is for Samples in Group (i.e. all samples from Condition1 followed by all samples from Condition2) (follow example_matrix_p1)) and P is for samples in Pairs (follow example_matrix_p4). By default it considers sample pairs in Group (G)")}
}
\value{
For each pair in the input matrix, it will return an output data-frame with w_positive, w_negative, P_value, and adjusted P_value for each sncRNA family chosen for analysis.
}
\description{
FUSION_ps - Enrichment Analysis of sncRNA families in paired data samples using expression matrix
}
\examples{
\donttest{
# Note:  To run and check the examples, please first set the working directory to the base folder of installed FUSION package. such as : setwd("/home/..../R/.../FUSION/").
FUSION_ps(a = "./extdata/example_matrix_p1.txt") ## Run enrichment analysis for example_matrix_p1.txt (5 pairs of samples) at default row_mean threshold (i.e., 0.1) for sncRNA families (tsRNAs, rsRNAs and ysRNAs) (default);
FUSION_ps(a = "./extdata/example_matrix_p1.txt", padj_method = "BH") ## Run enrichment analysis for example_matrix_p1.txt (5 pairs of samples) at default row_mean threshold (i.e., 0.1) for sncRNA families (tsRNAs, rsRNAs and ysRNAs) (default) using BH (Benjamini & Hochberg) method for correcting or adjusting p-values;
FUSION_ps(a = "./extdata/example_matrix_p1.txt", row_mean = 0.5) ## Run enrichment analysis for example_matrix_p1.txt (5 pairs of samples) at row_mean threshold of 0.5 for sncRNA families (tsRNAs, rsRNAs and ysRNAs) (default);
FUSION_ps(a = "./extdata/example_matrix_p2.txt", sncrna_family = "a")    ## Run enrichment analysis for example_matrix_p2.txt (10 pairs of samples) at default row_mean threshold (i.e., 0.1) for all sncRNA families;
FUSION_ps(a = "./extdata/example_matrix_p2.txt", sncrna_family = 0)    ## Run enrichment analysis for example_matrix_p2.txt (10 pairs of samples) at default row_mean threshold (i.e., 0.1) for all sncRNA families;
FUSION_ps(a = "./extdata/example_matrix_p2.txt", sncrna_family = "mirna")  ## Run enrichment analysis for example_matrix_p2.txt (10 pairs of samples) at default row_mean threshold (i.e., 0.1) for miRNA families;
FUSION_ps(a = "./extdata/example_matrix_p2.txt", sncrna_family = "rsrna")  ## Run enrichment analysis for example_matrix_p2.txt (10 pairs of samples) at default row_mean threshold (i.e., 0.1) for rsRNA families;
FUSION_ps(a = "./extdata/example_matrix_p3.txt", row_mean = 0.5, sncrna_family = "tsrna")   ## Run enrichment analysis for example_matrix_p2.txt (4 pairs of samples) at row_mean threshold of 0.5 for tsRNA families;
FUSION_ps(a = "./extdata/example_matrix_p3.txt", row_mean = 0.1, sncrna_family = "ysrna")  ## Run enrichment analysis for example_matrix_p2.txt (4 pairs of samples) at row_mean threshold of 0.1 for for ysRNA families;
FUSION_ps(a = "./extdata/example_matrix_p3.txt", sncrna_family = "other")  ## Run enrichment analysis for example_matrix_p2.txt (4 pairs of samples) at default row_mean threshold (i.e., 0.1) for other (pRNA,snRNA and snoRNA) sncRNA families;
FUSION_ps(a = "./extdata/example_matrix_p4.txt",order = "P") ## Run enrichment analysis for example_matrix_p4.txt (5 pairs of samples) at default row_mean threshold (i.e., 0.1) for sncRNA families (tsRNAs, rsRNAs and ysRNAs) (default) and samples are arranged in as pairs of columns;
# Note:  If you want to save the terminal/console output to a file, use sink() command. e.g., options(max.print = 1e6); sink("~/output.txt"); FUSION_ps(a = "./extdata/example_matrix_p1.txt"); sink()
}
}
