% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEretic.r
\name{getEretic}
\alias{getEretic}
\title{extract xml quantification information from a bruker xml}
\usage{
getEretic(path)
}
\arguments{
\item{path}{\itemize{
\item the path to the expName folder
}}
}
\value{
the title
}
\description{
extract xml quantification information from a bruker xml
}
