% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEreticFactor.r
\name{getEreticFactor}
\alias{getEreticFactor}
\title{extract eretic factor from eretic data frame}
\usage{
getEreticFactor(eretic)
}
\arguments{
\item{eretic}{\itemize{
\item eretic data frame
}}
}
\value{
the eretic factor
}
\description{
extract eretic factor from eretic data frame
}
