% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getID.R
\name{getID}
\alias{getID}
\title{S4 method get sampleID from dataElement}
\usage{
getID(da, using = "sampleID")
}
\arguments{
\item{da}{dataElement}

\item{using}{the name of the column used as UID}
}
\value{
the sampleID
}
\description{
S4 method get sampleID from dataElement
}
