% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfoFromExpNameList.R
\name{getInfoFromExpNameList}
\alias{getInfoFromExpNameList}
\title{get information from a list of runs}
\usage{
getInfoFromExpNameList(listOfExpName)
}
\arguments{
\item{listOfExpName}{\itemize{
\item a list of experiment folders
}}
}
\value{
a data frame with information
}
\description{
get information from a list of runs
}
