% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIvdr.r
\name{getIvdr}
\alias{getIvdr}
\title{extract small molecules  quantification information from a bruker xml}
\usage{
getIvdr(path, reportName)
}
\arguments{
\item{path}{\itemize{
\item the path to the expName folder
}}

\item{reportName}{\itemize{
\item name of the file (as it may change)
}}
}
\value{
the values
}
\description{
extract small molecules  quantification information from a bruker xml
}
