% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLipoReport.R
\name{getLipoTable}
\alias{getLipoTable}
\title{get names and description of lipoproteins IVDr parameters}
\usage{
getLipoTable()
}
\value{
a data.frame with information
}
\description{
get names and description of lipoproteins IVDr parameters
}
