% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParam.R
\name{getParam}
\alias{getParam}
\title{get a parameter from procs or acqus dataframe}
\usage{
getParam(df, paramName)
}
\arguments{
\item{df}{\itemize{
\item the name of the dataframe
}}

\item{paramName}{\itemize{
\item the name of the parameter to get
}}
}
\value{
the parameter
}
\description{
get a parameter from procs or acqus dataframe
}
