% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQcReport.r
\name{getQcReport}
\alias{getQcReport}
\title{extract lipoprotein  quantification information from a bruker xml}
\usage{
getQcReport(path, reportName)
}
\arguments{
\item{path}{\itemize{
\item the path to the expName folder
}}

\item{reportName}{\itemize{
\item the file name of the report
(plamsa_qc_report.xml / urine_qc_report.xml)
}}
}
\value{
the values
}
\description{
extract lipoprotein  quantification information from a bruker xml
}
