% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpectra.R
\name{getSpectra}
\alias{getSpectra}
\title{extract a spectrum (processed) from a folder}
\usage{
getSpectra(path, expName, expNo)
}
\arguments{
\item{path}{\itemize{
\item the path to the expName folder
}}

\item{expName}{\itemize{
\item the name of the folder with experiments
}}

\item{expNo}{\itemize{
\item the number of experiment folder
}}
}
\value{
a vector with spectra (real part and x axis)
}
\description{
extract a spectrum (processed) from a folder
}
