% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTitle.R
\name{getTitle}
\alias{getTitle}
\title{extract title from a bruker folder}
\usage{
getTitle(filePath)
}
\arguments{
\item{filePath}{\itemize{
\item the path to the expName folder
}}
}
\value{
the title
}
\description{
extract title from a bruker folder
}
