% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getType.R
\name{getType}
\alias{getType}
\title{S4 method get sampleType from dataElement}
\usage{
getType(da, using = "sampleType")
}
\arguments{
\item{da}{dataElement}

\item{using}{the name of the column used as sampleType}
}
\value{
the sampleType
}
\description{
S4 method get sampleType from dataElement
}
