% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftJoinWithAcqus.R
\name{leftJoinWithAcqus}
\alias{leftJoinWithAcqus}
\title{function to left join data to acqus}
\usage{
leftJoinWithAcqus(A, B, by)
}
\arguments{
\item{A}{\itemize{
\item first data.frame (acqus)
}}

\item{B}{\itemize{
\item second data.frame (qc)
}}

\item{by}{\itemize{
\item the name of the key column
}}
}
\value{
a data.frame with qc matched to acqus
}
\description{
function to left join data to acqus
}
