% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipo-data.r
\docType{data}
\name{lipo}
\alias{lipo}
\title{Lipoproteins from IVDr method}
\format{
A data frame with 112 lipoproteins values
}
\usage{
data(lipo)
}
\description{
A dataset containing 112 Lipoproteins
}
\details{
\itemize{
\item Fraction - lipoprotein fraction
}
}
\keyword{datasets}
