% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.dataElement.R
\name{load.daE}
\alias{load.daE}
\title{A nrow method for \emph{dataElement} objects}
\usage{
load.daE(file)
}
\arguments{
\item{file}{a path to the file to load}
}
\value{
the content of the dataElement
}
\description{
A nrow method for \emph{dataElement} objects
}
