% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.dataElement.R
\name{loadAs}
\alias{loadAs}
\title{A nrow method for \emph{dataElement} objects}
\usage{
loadAs(filepath)
}
\arguments{
\item{filepath}{a path to the file to load}

\item{as}{name of the variable where to store the content}

\item{...}{parameters of the nrow generic function}
}
\value{
the content of the dataElement
}
\description{
A nrow method for \emph{dataElement} objects
}
