% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup}
\alias{lookup}
\title{function to lookup annotations for each samples}
\usage{
lookup(da, ann, using = c("sourceID", "sampleID"))
}
\arguments{
\item{da}{a dataElement}

\item{ann}{a data.frame with annotations}

\item{using}{the name of the column to lookup}
}
\value{
a dataElement with matched samples
}
\description{
function to lookup annotations for each samples
}
