% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLipoReport.R
\name{makeLipoReport}
\alias{makeLipoReport}
\title{print lipoprotein  quantification report}
\usage{
makeLipoReport(lip, options = list())
}
\arguments{
\item{lip}{\itemize{
\item a lipoprotein report
}}

\item{options}{\itemize{
\item a list of options
}}

\item{options$fold}{\itemize{
\item if true use design for fold change (default = FALSE)
}}

\item{options$scale}{\itemize{
\item if true center and scale (align) the output (default = FALSE)
}}
}
\value{
print a report
}
\description{
print lipoprotein  quantification report
}
