% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSecret.R
\name{makeSecret}
\alias{makeSecret}
\title{function to make string secret to protect data privacy}
\usage{
makeSecret(string)
}
\arguments{
\item{string}{string}
}
\value{
md5sum of the string
}
\description{
function to make string secret to protect data privacy
}
