% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSPC.R
\name{makeSPC}
\alias{makeSPC}
\title{function to make dataElements of SPC 1,2,3 and Glyc A and B integrals}
\usage{
makeSPC(da, calibrate = TRUE)
}
\arguments{
\item{da}{A data element.}

\item{calibrate}{logical}
}
\value{
A data element with SPC and Glyc integrals
}
\description{
function to make dataElements of SPC 1,2,3 and Glyc A and B integrals
}
\examples{
# da<-local(get(load(~.daE)))
# test<-makeSPC(da,calibrate = TRUE)

}
