% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUnique.R
\name{makeUnique}
\alias{makeUnique}
\title{make a vector of character unique}
\usage{
makeUnique(sampleID)
}
\arguments{
\item{sampleID}{\itemize{
\item a vector of characters
}}
}
\value{
a vector of unique characters
}
\description{
matching different dataElements is only possible if all possess
a vector with unique IDs, sampleID in its experimentalParameters.
If this vector is not unique, due to the presence of quality
controls or reference samples, then this function makes it unique
by numbering all duplicates
}
