% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchAll.R
\name{matchAll}
\alias{matchAll}
\title{function to match all dataElements}
\usage{
matchAll(list)
}
\arguments{
\item{list}{list of dataElements}
}
\value{
all dataElements matched by sampleID
}
\description{
function to match all dataElements
}
