% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPairwise.R
\name{matchPairwise}
\alias{matchPairwise}
\title{function to match two dataElement}
\usage{
matchPairwise(daA, daB)
}
\arguments{
\item{daA}{first dataElement}

\item{daB}{second dataElement}
}
\value{
two dataElement with matched samples
}
\description{
function to match two dataElement
}
