% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchWith.R
\name{matchWith}
\alias{matchWith}
\title{function to match one dataElement against a second one}
\usage{
matchWith(daA, daB)
}
\arguments{
\item{daA}{first dataElement}

\item{daB}{second dataElement}
}
\value{
dataElement with matched samples
}
\description{
function to match one dataElement against a second one
}
