% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRClassDefinitions.R
\docType{class}
\name{NMRPeak1D-class}
\alias{NMRPeak1D-class}
\title{An S4 class for NMRPeak1D}
\value{
a dataElement
}
\description{
An S4 class for NMRPeak1D
}
\section{Slots}{

\describe{
\item{\code{x}}{Ordinate of the peak}

\item{\code{y}}{height of the peak}

\item{\code{fwhm}}{Full width at half maximum of peak}

\item{\code{shape}}{(optional) A peak shape for the different peaks conforming the signal. It uses the parent shape if void}

\item{\code{type}}{The Class Name. Used for moving between S4 and JSON}
}}

