% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRClassDefinitions.R
\docType{class}
\name{NMRSignal1D-class}
\alias{NMRSignal1D-class}
\title{An S4 class for NMRSignal1D}
\value{
a dataElement
}
\description{
An S4 class for NMRSignal1D
}
\section{Slots}{

\describe{
\item{\code{peaks}}{a list of NMRPeak1D}

\item{\code{nbAtoms}}{Number of atoms associated with the signal}

\item{\code{integration}}{Raw non-normalized integration of the signal}

\item{\code{chemicalShift}}{(optional) The chemical shift of the signal. Not always its center}

\item{\code{multiplicity}}{(optional) A compiled NMR multiplicity pattern i.e: s|d|t|q|s,...|dd,...}

\item{\code{shiftRange}}{(optional) Range of x-peaks variation. It is an absolute value. Should be positive}

\item{\code{heightRangePer}}{(optional) Proportional range of y-peaks variation. Must be between 0 and 1}

\item{\code{widthFactor}}{(optional) Width factor depending on the signal}

\item{\code{shape}}{(optional) A peak shape for the different peaks conforming the signal. Internal components overrides this shape}

\item{\code{diaIDs}}{(optional) A list of atom ids to which this signal is assigned.}

\item{\code{analyte}}{(optional) The name/id of the analyte}

\item{\code{validated}}{(optional) Validated by an expert?}

\item{\code{type}}{The Class Name. Used for moving between S4 and JSON}

\item{\code{id}}{(optional) Unique identifier}
}}

