% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRClassDefinitions.R
\docType{class}
\name{NMRSignalModel-class}
\alias{NMRSignalModel-class}
\title{An S4 class for NMRSignalModel}
\value{
a dataElement
}
\description{
An S4 class for NMRSignalModel
}
\section{Slots}{

\describe{
\item{\code{signals}}{a list of NMRSignal1D}

\item{\code{from}}{start point for ROI}

\item{\code{to}}{end point for ROI}

\item{\code{ppm}}{Array of x-values from spectrum}

\item{\code{experimental}}{Array of y-values from spectrum}

\item{\code{fitted}}{Array of y-values given by the optimization}

\item{\code{signalsOutput}}{A list of signal inputs with the optimized parameters}

\item{\code{shape}}{(optional) A peak shape for the different peaks conforming the signal. Internal components overrides this shape}

\item{\code{error}}{A list of different errors. I'll explain later}

\item{\code{id}}{(optional) Unique identifier}
}}

