% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow.dataElement.R
\name{nrow.dataElement}
\alias{nrow.dataElement}
\title{S3 method nrow for dataElement objects (S4)}
\usage{
nrow.dataElement(da, ...)
}
\arguments{
\item{da}{a dataElement}

\item{...}{parameters of the nrow generic function}
}
\value{
the number of rows
}
\description{
A nrow method for \emph{dataElement} objects
}
