% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numToSignificance.R
\name{numToSignificance}
\alias{numToSignificance}
\title{convert values to significance}
\usage{
numToSignificance(val)
}
\arguments{
\item{val}{\itemize{
\item a vector of values
}}
}
\value{
a vector with significance text
}
\description{
convert values to significance
}
