% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{order}
\alias{order}
\title{S4 method to reorder a dataElement}
\usage{
order(da, idx)
}
\arguments{
\item{idx}{the index}

\item{daA}{first dataElement}
}
\value{
the reordered dataElement
}
\description{
S4 method to reorder a dataElement
}
