% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMS.R
\name{parseMS}
\alias{parseMS}
\title{parseMS}
\usage{
parseMS(folder, assay, fileType = NULL, optns = list())
}
\arguments{
\item{folder}{Either character (file path) to folder/file or a data frame from a mass spec rolodex.}

\item{assay}{character for mass spec assay. Must be one of "AA", "TRY", "BA", "SCFA".}

\item{fileType}{character for file type if folder argument is a path to a folder/file.
Must be "txt", "tsv" or "xml".}

\item{optns}{List for extra arguments}
}
\value{
data element
}
\description{
parseMS
}
