% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTargetedMS.r
\name{parseTargetedMS}
\alias{parseTargetedMS}
\title{importation function for targeted MS assays}
\usage{
parseTargetedMS(file, method)
}
\arguments{
\item{file}{\itemize{
\item the file path to be imported
}}

\item{method}{\itemize{
\item the method used to acquire the assay
}}
}
\value{
a dataElement
}
\description{
importation function for targeted MS assays
}
