% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{plot correlation between NMR trace (part of) and MS data}
\usage{
plotCorrelation(x, labels, trace, xaxis, options = list())
}
\arguments{
\item{x}{\itemize{
\item a matrix with correlation values
}}

\item{labels}{\itemize{
\item the name of the MS values (x rows)
}}

\item{trace}{\itemize{
\item the NMR trace to plot (values)
}}

\item{xaxis}{\itemize{
\item the xaxis for the NMR trace
}}

\item{options}{\itemize{
\item options
}}
}
\description{
plot correlation between NMR trace (part of) and MS data
}
