% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{plotMap}
\alias{plotMap}
\title{plot a geographical map}
\usage{
plotMap(
  data,
  latitude = c(-45, 15),
  longitude = c(-150, 180),
  colorRange = c("#ffeeee", "#ff0000"),
  breaks = c(100, 200, 500, 1000, 3000)
)
}
\arguments{
\item{latitude}{\itemize{
\item the latitude range
}}

\item{longitude}{\itemize{
\item the longitude range
}}

\item{colorRange}{\itemize{
\item color range
}}

\item{breaks}{\itemize{
\item breaks for color range
}}

\item{values}{\itemize{
\item a data frame with countries
}}
}
\value{
print a map
}
\description{
plot a geographical map
}
