% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSHY.r
\name{plotSHY}
\alias{plotSHY}
\title{plot correlation between NMR trace (part of) and MS data with segments}
\usage{
plotSHY(cor, sig, txt, labels, trace, xaxis, options = list())
}
\arguments{
\item{cor}{\itemize{
\item a matrix with correlation values
}}

\item{sig}{\itemize{
\item a matrix with significance values
}}

\item{txt}{\itemize{
\item a matrix with text
}}

\item{labels}{\itemize{
\item the name of the MS values (x rows)
}}

\item{trace}{\itemize{
\item the NMR trace to plot (values)
}}

\item{xaxis}{\itemize{
\item the xaxis for the NMR trace
}}

\item{options}{\itemize{
\item options
}}
}
\description{
plot correlation between NMR trace (part of) and MS data with segments
}
