% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSHY2.r
\name{plotSHY2}
\alias{plotSHY2}
\title{plot correlation between NMR trace (part of) and MS data with segments}
\usage{
plotSHY2(corList, sigList, txtList, labels, trace, xaxis, options = list())
}
\arguments{
\item{corList}{\itemize{
\item a list with correlation values
}}

\item{sigList}{\itemize{
\item a list with significance values
}}

\item{txtList}{\itemize{
\item a List with text
}}

\item{labels}{\itemize{
\item the name of the MS values (x rows)
}}

\item{trace}{\itemize{
\item a list of NMR trace to plot (values)
}}

\item{xaxis}{\itemize{
\item the xaxis for the NMR trace
}}

\item{options}{\itemize{
\item options
}}
}
\description{
plot correlation between NMR trace (part of) and MS data with segments
}
