% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dataElement.R
\name{print.dataElement}
\alias{print.dataElement}
\title{print information for dataElement objects}
\usage{
\method{print}{dataElement}(x, ...)
}
\arguments{
\item{x}{dataElement}

\item{...}{other parameters}
}
\value{
the information about dataElement
}
\description{
print information for dataElement objects
}
