% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLipoReport.R
\name{printCaption}
\alias{printCaption}
\title{print caption for lipoprotein report}
\usage{
printCaption(caption, row, column)
}
\arguments{
\item{caption}{\itemize{
\item the caption
}}

\item{row}{\itemize{
\item row value
}}

\item{column}{\itemize{
\item column value
}}
}
\description{
print caption for lipoprotein report
}
