% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLipoReport.R
\name{printIndicator}
\alias{printIndicator}
\title{print lipoprotein indicator in report}
\usage{
printIndicator(lip, row, column, options = list())
}
\arguments{
\item{lip}{\itemize{
\item indicator value
}}

\item{row}{\itemize{
\item row value
}}

\item{column}{\itemize{
\item column value
}}

\item{options}{= a list with options}
}
\description{
print lipoprotein indicator in report
}
