% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLipoReport.R
\name{printStrip}
\alias{printStrip}
\title{print strip in lipoprotein report}
\usage{
printStrip(lip, row, column, options = list())
}
\arguments{
\item{lip}{\itemize{
\item lipo value
}}

\item{row}{\itemize{
\item row value
}}

\item{column}{\itemize{
\item column value
}}

\item{options}{\itemize{
\item a list of options
}}
}
\description{
print strip in lipoprotein report
}
