% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLipoReport.R
\name{printTitle}
\alias{printTitle}
\title{print title for lipoprotein report}
\usage{
printTitle(title, titleBoxPosition, titlePosition)
}
\arguments{
\item{title}{\itemize{
\item the title
}}

\item{titleBoxPosition}{\itemize{
\item position
}}

\item{titlePosition}{\itemize{
\item position of the title in the box
}}
}
\description{
print title for lipoprotein report
}
