% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExperiment.r
\name{readExperiment}
\alias{readExperiment}
\title{read experiment from a bruker folder (expno)}
\usage{
readExperiment(path, acqus = TRUE, procs = TRUE)
}
\arguments{
\item{path}{\itemize{
\item the path to the expNo folder
}}

\item{acqus}{\itemize{
\item object with acquisiton parameters, if TRUE reads from the file
}}

\item{procs}{\itemize{
\item object with processing parameters, if TRUE reads from the file
}}
}
\value{
a list with all read elements
}
\description{
read experiment from a bruker folder (expno)
}
