% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readParam.r
\name{readParam}
\alias{readParam}
\title{extract a parameter from a bruker file (procs or acqus)}
\usage{
readParam(path, paramName)
}
\arguments{
\item{path}{\itemize{
\item the path to the expName folder
}}

\item{paramName}{\itemize{
\item the name of the parameter to read
}}
}
\value{
the parameter
}
\description{
extract a parameter from a bruker file (procs or acqus)
}
