% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXML.r
\name{readXML}
\alias{readXML}
\title{extract content from a xml object}
\usage{
readXML(xml, attr = FALSE)
}
\arguments{
\item{xml}{\itemize{
\item an xml object
}}

\item{attr}{\itemize{
\item boolean object
}}
}
\value{
assign value to content
}
\description{
extract content from a xml object
}
