% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectWith.R
\name{selectWith,dataElement,numeric-method}
\alias{selectWith,dataElement,numeric-method}
\title{S4 method to filter a dataElement}
\usage{
\S4method{selectWith}{dataElement,numeric}(da, idx)
}
\arguments{
\item{da}{first dataElement}

\item{idx}{the index for selection}
}
\value{
the reordered dataElement
}
\description{
S4 method to filter a dataElement
}
