% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectWith.R
\name{selectWith}
\alias{selectWith}
\title{S4 method to select from a dataElement}
\usage{
selectWith(da, idx)
}
\arguments{
\item{da}{first dataElement}

\item{idx}{the filter}
}
\value{
the selected row of dataElement
}
\description{
S4 method to select from a dataElement
}
