% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setID.R
\name{setID,dataElement-method}
\alias{setID,dataElement-method}
\title{S4 method set sampleID from dataElement}
\usage{
\S4method{setID}{dataElement}(da, ID, using = "sampleID")
}
\arguments{
\item{da}{dataElement}

\item{ID}{a vector of IDs}

\item{using}{the name of the column used as UID}
}
\value{
void
}
\description{
S4 method set sampleID from dataElement
}
