% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showDuplicated}
\alias{showDuplicated}
\title{a function to show pairs of duplicated items}
\usage{
showDuplicated(aTable, colWithDuplicates, colToSelect)
}
\arguments{
\item{aTable}{\itemize{
\item a table
}}

\item{colWithDuplicates}{= the column with duplicates}

\item{colToSelect}{\itemize{
\item the column to display
}}
}
\value{
a logical vector
}
\description{
a function to show pairs of duplicated items
}
