% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp}
\alias{stamp}
\title{get time stamp to the millisecond to use as uuid generator when no uuid is
available via rolodex database.}
\usage{
stamp(n)
}
\arguments{
\item{n}{number of ids}
}
\value{
a number that is hopefully unique
}
\description{
get time stamp to the millisecond to use as uuid generator when no uuid is
available via rolodex database.
}
