% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRClassDefinitions.R
\name{toJSONFile,character-method}
\alias{toJSONFile,character-method}
\title{Method for creating a JSON file out of a character}
\usage{
\S4method{toJSONFile}{character}(obj, control = NA, con = "ANY")
}
\arguments{
\item{obj}{A data object to be parsed (list, array or S4)}

\item{control}{A set of control parameters for the transformation. We use it to avoid export some S4 slots}

\item{con}{A connection to the output file}
}
\value{
void
}
\description{
Method for creating a JSON file out of a character
}
