% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncalibrate.R
\name{unclibrate}
\alias{unclibrate}
\title{function to reset spectral reference to zero}
\usage{
unclibrate(data)
}
\arguments{
\item{data}{\itemize{
\item dataElements
}}
}
\value{
dataElements
}
\description{
function to reset spectral reference to zero
}
\examples{
# dae<-local(get(load(~.daE)))
#test<-unclibrate(data = dae)

}
