% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRClassDefinitions.R
\name{writeToJSON}
\alias{writeToJSON}
\title{Create and simplify a JSON file out of a data object.}
\usage{
writeToJSON(data, fileName)
}
\arguments{
\item{data}{A data object to be parsed (list, array or S4)}

\item{fileName}{The name of the file for storing the result}
}
\value{
void
}
\description{
Create and simplify a JSON file out of a data object.
}
