% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{S4 method to check dataElement
#'}
\usage{
check(da, using = "sampleID")
}
\arguments{
\item{da}{dataElement object to be checked}
}
\value{
void

check("dataElement")
}
\description{
S4 method to check dataElement
#'
}
