% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSION.R
\name{FUSION}
\alias{FUSION}
\title{FUSION  - Family-level Unique Small RNA IntegratION across samples using expression matrix}
\arguments{
\item{a}{a matrix with first and second column as "Sequence" (or ID) and "Annotation", respectively. Sequence (or ID) must be unique.  Rest of the columns provides RPM or other normalized expression values from different samples under study}

\item{row_mean}{mean RPM (default 0.1) threshold to retain the sncRNA species (rows) in the matrix}

\item{sncrna_family}{list of sncRNA families to be analyzed in the abundance analysis study. Use "tsrna", "rsrna", "ysrna", or "mirna" for tsRNAs, rsRNAs, ysRNAs, or mirna, respectively. Use "other" for 'pRNA, snRNA and snoRNA'. For all, use any letter or number, eg. "a","b", "c", 1, 2, 3. By default (i.e. if no option specified) it will search for tryRNAs (tsRNAs, rsRNAs, and ysRNAs)}
}
\value{
with FUSION_ps, it will return, for each pair in the input matrix, an output data-frame with w_positive, w_negative, P_value, and adjusted P_value for each sncRNA family chosen for analysis.

with FUSION_ms, it will return a final output in a data-frame with t-statistics, P_value, and adjusted P_value for each sncRNA family chosen for analysis

with FUSION_msmc, it will return a final output in a data-frame with t-statistics, P_value, and adjusted P_value for each sncRNA family chosen for analysis
}
\description{
FUSION  - Family-level Unique Small RNA IntegratION across samples using expression matrix
}
\details{
For more information, see the user manual:
\code{browseURL(system.file("docs/FUSION_v1.0.2_User_Manual.pdf", package = "FUSION"))}

After installing the package FUSION, call the library as: library(FUSION).
\cr
\cr There are three main functions in this package : FUSION_ps, FSION_ms, and FUSION_msmc
\cr
\cr Ask help for the description and help menus of each of these functions :
\cr ?FUSION_ps
\cr ?FUSION_ms
\cr ?FUSION_msmc
\cr
\cr If want to run examples, it is necessary to set working directory to the base folder of the installed package FUSION. Such as : setwd("/home/..../R/.../FUSION/").
\cr
}
