% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{fusion}
\alias{fusion}
\title{function to match two dataElement}
\usage{
fusion(daA, daB)
}
\arguments{
\item{daA}{first dataElement}

\item{daB}{second dataElement}
}
\value{
two dataElement with matched samples
}
\description{
function to match two dataElement
}
