% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData,dataElement,ANY,character-method}
\alias{getData,dataElement,ANY,character-method}
\title{S4 method get data from dataElement}
\usage{
\S4method{getData}{dataElement,ANY,character}(da, type = c("sample"), replicates = "last")
}
\arguments{
\item{da}{dataElement}

\item{using}{either sample, qc, standard, or blank or a combination thereof}
}
\value{
a dataElement with selected types
}
\description{
S4 method get data from dataElement
}
