% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{S4 method get data from dataElement}
\usage{
getData(da, type)
}
\arguments{
\item{da}{dataElement}

\item{using}{either sample, qc, standard, or blank or a combination of}
}
\value{
a dataElement with selected types
}
\description{
S4 method get data from dataElement
}
