% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getID.R
\name{getID,dataElement-method}
\alias{getID,dataElement-method}
\title{S4 method get sampleID from dataElement}
\usage{
\S4method{getID}{dataElement}(da, using = "sampleID", re.rm = FALSE)
}
\arguments{
\item{da}{dataElement}

\item{using}{the name of the column used as UID}
}
\value{
the sampleID
}
\description{
S4 method get sampleID from dataElement
}
