% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getID.R
\name{getID}
\alias{getID}
\title{S4 method get sampleID from dataElement}
\usage{
getID(da, using = "sampleID", re.rm = FALSE)
}
\arguments{
\item{da}{dataElement}

\item{using}{the name of the column used as UID}

\item{re.rm}{if TRUE replicate markers are removed}
}
\value{
the sampleID
}
\description{
S4 method get sampleID from dataElement
}
