% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParam.R
\name{getParam}
\alias{getParam}
\title{extract a parameter from a bruker folder}
\usage{
getParam(filePath, paramName)
}
\arguments{
\item{filePath}{\itemize{
\item the path to the expName folder
}}

\item{paramName}{\itemize{
\item the name of the parameter to read
}}
}
\value{
the parameter
}
\description{
extract a parameter from a bruker folder
}
