% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTRY.r
\name{getTRY}
\alias{getTRY}
\title{extract targeted MS data from task export xml file}
\usage{
getTRY(path)
}
\arguments{
\item{path}{\itemize{
\item the path to the expName folder
}}
}
\value{
a dataElement object
}
\description{
extract targeted MS data from task export xml file
}
