% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUnique.R
\name{makeUnique}
\alias{makeUnique}
\title{function to make ID unique}
\usage{
makeUnique(sampleID, sep = "#", fromFirst = FALSE, reverse = FALSE, first = 1)
}
\arguments{
\item{sampleID}{sampleID}

\item{sep}{the separator}

\item{fromFirst}{if true all replicate including the first will be marked}

\item{reverse}{if true start marking from last}

\item{first}{a number to start numbering from}
}
\value{
unique ID
}
\description{
function to make ID unique
}
\examples{
a <- c("a", "a", "b", "a", "b", "c")
makeUnique(a)

}
