% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUnique.R
\name{makeUnique}
\alias{makeUnique}
\title{function to make ID unique}
\usage{
makeUnique(sampleID, sep = "#")
}
\arguments{
\item{sep}{the separator}

\item{da}{sampleID}
}
\value{
unique ID
}
\description{
function to make ID unique
}
