% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchAll.R
\name{matchAll}
\alias{matchAll}
\title{function to match all dataElement}
\usage{
matchAll(list)
}
\arguments{
\item{list}{list of dataElement}
}
\value{
all dataElement matched by sampleID
}
\description{
function to match all dataElement
}
