% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchTo.R
\name{matchTo}
\alias{matchTo}
\title{function to match one dataElement against a second one}
\usage{
matchTo(daA, daB, using = "sampleID")
}
\arguments{
\item{daA}{first dataElement}

\item{daB}{second dataElement}

\item{using}{the name of the column used as UID}
}
\value{
dataElement with matched samples
}
\description{
function to match one dataElement against a second one
}
