% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderWith.R
\name{orderWith,dataElement,numeric-method}
\alias{orderWith,dataElement,numeric-method}
\title{S4 method to reorder a dataElement}
\usage{
\S4method{orderWith}{dataElement,numeric}(da, idx)
}
\arguments{
\item{idx}{the index}

\item{daA}{first dataElement}
}
\value{
the reordered dataElement
}
\description{
S4 method to reorder a dataElement
}
