% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderWith.R
\name{orderWith}
\alias{orderWith}
\title{S4 method to reorder a dataElement}
\usage{
orderWith(da, idx)
}
\arguments{
\item{idx}{the index}

\item{daA}{first dataElement}
}
\value{
the reordered dataElement
}
\description{
a dataElement is ordered using the index provided as second
argument. This is to ensure that both the .Data and the
experimentalParameters slots are rotated accordingly
}
