% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseNMR.R
\name{parseNMR}
\alias{parseNMR}
\title{reads all information from bruker folder structures}
\usage{
parseNMR(folder, opts = NULL)
}
\arguments{
\item{folder}{\itemize{
\item the root folder to read from
}}

\item{opts}{\itemize{
\item opts (what(spec, brxlipo, prxpacs, brxsm, spcglyc), specOpts(procno, uncalibrate, fromTo, length.out))
}}
}
\description{
reads all information from bruker folder structures
}
