% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTargetedMS.r
\name{parseTargetedMS}
\alias{parseTargetedMS}
\title{importation function for targeted MS assays}
\usage{
parseTargetedMS(file, method, options = list())
}
\arguments{
\item{file}{\itemize{
\item the file path to be imported
}}

\item{method}{\itemize{
\item the method used to acquire the assay
}}

\item{options}{list of options
\itemize{
\item codePosition - position of the code in the file name
\item columnList - list of columns to be selected
}}
}
\value{
a dataElement
}
\description{
importation function for targeted MS assays
}
