% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.r
\name{rbind.dataElement}
\alias{rbind.dataElement}
\title{rbind dataElements}
\usage{
\method{rbind}{dataElement}(...)
}
\arguments{
\item{...}{a list of dataElements}
}
\value{
the information about dataElement
}
\description{
rbind dataElements
}
