% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.r
\name{rbind.dataElement}
\alias{rbind.dataElement}
\title{print information for dataElement objects}
\usage{
\method{rbind}{dataElement}(...)
}
\value{
the information about dataElement
}
\description{
print information for dataElement objects
}
