% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.r
\name{rbind.dataElement}
\alias{rbind.dataElement}
\title{Bind together multiple dataElement objects by rows.}
\usage{
\method{rbind}{dataElement}(...)
}
\arguments{
\item{...}{a list of dataElements}
}
\value{
the information about dataElement
}
\description{
This function takes a list of dataElement objects and binds them together
by rows into a single dataElement object. It checks that the dataElements
are compatible in terms of type and method before binding. The observation
descriptions are concatenated and made unique where necessary.
rbind dataElements
}
