% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExperiment.r
\name{readExperiment}
\alias{readExperiment}
\title{read experiment from a bruker folder (expno)}
\usage{
readExperiment(path)
}
\arguments{
\item{path}{\itemize{
\item the path to the expNo folder
}}
}
\value{
a list with all read elements
}
\description{
read experiment from a bruker folder (expno)
}
