% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSpectrum.R
\name{readSpectrum}
\alias{readSpectrum}
\title{read a spectrum (processed) from a Bruker expno folder}
\usage{
readSpectrum(path, procs = TRUE, options = list())
}
\arguments{
\item{path}{\itemize{
\item the path to the expNo folder
}}

\item{procs}{\itemize{
\item the name of the folder with experiments
}}

\item{options}{\itemize{
\item options
}}

\item{options$uncalibrate}{\itemize{
\item if true, calibration will be removed (SR set to 0)
}}

\item{options$eretic}{\itemize{
\item provide eretic correction to be applied
}}

\item{options$fromTo}{\itemize{
\item provide lower and upper bound for common grid
}}

\item{options$length.out}{-set the length of the final data}
}
\value{
a vector with spectra (real part and x axis)
}
\description{
read a spectrum (processed) from a Bruker expno folder
}
